tests/test_text_operations.py
import unittest
from unittest.mock import patch, mock_open, MagicMock
import sys
import logging

# Import the function to be tested
from app.text_operations import process_text

class TestProcessText(unittest.TestCase):

    @patch('builtins.open', new_callable=mock_open)
    def test_file_not_found_error(self, mock_open):
        mock_open.side_effect = FileNotFoundError("File not found")
        with self.assertLogs(level='ERROR') as log:
            with self.assertRaises(SystemExit) as cm:
                process_text('nonexistent_file.txt', None)
            self.assertEqual(cm.exception.code, 1)
            self.assertIn("File not found", log.output[0])

    @patch('builtins.open', new_callable=mock_open)
    def test_io_error(self, mock_open):
        mock_open.side_effect = IOError("IO error")
        with self.assertLogs(level='ERROR') as log:
            with self.assertRaises(SystemExit) as cm:
                process_text('file.txt', None)
            self.assertEqual(cm.exception.code, 1)
            self.assertIn("IO error", log.output[0])

    @patch('app.file_operations.read_tokens')
    def test_value_error(self, mock_read_tokens):
        mock_read_tokens.side_effect = ValueError("Value error")
        with self.assertLogs(level='ERROR') as log:
            with self.assertRaises(SystemExit) as cm:
                process_text('file.txt', None)
            self.assertEqual(cm.exception.code, 1)
            self.assertIn("Value error", log.output[0])

    @patch('app.token_operations.detokenize')
    def test_key_error(self, mock_detokenize):
        mock_detokenize.side_effect = KeyError("Key error")
        with self.assertLogs(level='ERROR') as log:
            with self.assertRaises(SystemExit) as cm:
                process_text('file.txt', None)
            self.assertEqual(cm.exception.code, 1)
            self.assertIn("Key error", log.output[0])

    @patch('app.token_operations.detokenize')
    def test_type_error(self, mock_detokenize):
        mock_detokenize.side_effect = TypeError("Type error")
        with self.assertLogs(level='ERROR') as log:
            with self.assertRaises(SystemExit) as cm:
                process_text('file.txt', None)
            self.assertEqual(cm.exception.code, 1)
            self.assertIn("Type error", log.output[0])

    @patch('app.file_operations.read_tokens')
    def test_memory_error(self, mock_read_tokens):
        mock_read_tokens.side_effect = MemoryError("Memory error")
        with self.assertLogs(level='ERROR') as log:
            with self.assertRaises(SystemExit) as cm:
                process_text('file.txt', None)
            self.assertEqual(cm.exception.code, 1)
            self.assertIn("Memory error", log.output[0])

    @patch('builtins.open', new_callable=mock_open)
    def test_permission_error(self, mock_open):
        mock_open.side_effect = PermissionError("Permission error")
        with self.assertLogs(level='ERROR') as log:
            with self.assertRaises(SystemExit) as cm:
                process_text('file.txt', None)
            self.assertEqual(cm.exception.code, 1)
            self.assertIn("Permission error", log.output[0])

    @patch('builtins.open', new_callable=mock_open)
    def test_eof_error(self, mock_open):
        mock_open.side_effect = EOFError("EOF error")
        with self.assertLogs(level='ERROR') as log:
            with self.assertRaises(SystemExit) as cm:
                process_text('file.txt', None)
            self.assertEqual(cm.exception.code, 1)
            self.assertIn("EOF error", log.output[0])

    @patch('builtins.open', new_callable=mock_open)
    def test_unicode_decode_error(self, mock_open):
        mock_open.side_effect = UnicodeDecodeError("utf-8", b"", 0, 1, "Unicode decode error")
        with self.assertLogs(level='ERROR') as log:
            with self.assertRaises(SystemExit) as cm:
                process_text('file.txt', None)
            self.assertEqual(cm.exception.code, 1)
            self.assertIn("Unicode decode error", log.output[0])

    @patch('app.file_operations.read_tokens')
    def test_unexpected_error(self, mock_read_tokens):
        mock_read_tokens.side_effect = Exception("Unexpected error")
        with self.assertLogs(level='ERROR') as log:
            with self.assertRaises(SystemExit) as cm:
                process_text('file.txt', None)
            self.assertEqual(cm.exception.code, 1)
            self.assertIn("An unexpected error occurred", log.output[0])

if __name__ == '__main__':
    unittest.main()